fix: OfferPreviewModal .web file
